azure: report failure to host if ephemeral DHCP secondary NIC (#4558)
Azure requires the use of the primary interface to query IMDS
and report ready to wireserver.  If we determine that the interface
is not the primary interface, report the failure to the host.

The exception is when iterating over hot-plugged NICs in the Savable
PPS scenario.  In this case the secondary may be enumerated first
and we would expect a failure.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>