tests: hard-code canonical-kernel-team ppa to noble
Hard-code canonical-kernel-team PPA to most recent LTS to
avoid test failures on the curent development release due
to no signed Release file for oracular. A missing Release file
is an indication that no packages have yet been uploaded for
the development series.

Using a hard-coded noble in test_apt_functionality does not
affect the integrity of the apt_functionality test because this
PPA is only used to assert the valid GPG key is honored. No
packages are installed from the PPA so no issue arises with
package dependencies.